fix(ci): thermos round-2 hygiene findings
