+++ scripts/check_xpass.py
@@ -45,12 +45,12 @@ class XpassInventory(NamedTuple):
         return len(self.records)
 
     @property
-    def xpass_records(self) -> tuple[XpassRecord, ...]:
+    def failing_records(self) -> tuple[XpassRecord, ...]:
         """Return every parsed XPASS line (all are gate failures)."""
         return self.records
 
     @property
-    def xpass_count(self) -> int:
+    def failing_count(self) -> int:
         """Return the number of XPASS lines (the fail condition)."""
         return len(self.records)
 
@@ -89,8 +89,8 @@ def check_xpass(inventory: XpassInventory, *, stream: TextIO | None = None) -> i
         Process exit code (0 ok, 1 xpasses remain).
     """
     out: TextIO = sys.stderr if stream is None else stream
-    records = inventory.xpass_records
-    summary = f"{inventory.xpass_count} xpassed ({inventory.total} total)"
+    records = inventory.failing_records
+    summary = f"{inventory.failing_count} xpassed ({inventory.total} total)"
     if not records:
         print(f"xpass-check OK: {summary}", file=out)
         return 0
